+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [0.0.27](https://github.com/juztadrop/juztadrop/compare/v0.0.26...v0.0.27) (2026-03-11)
+
+### Changes
+
+- feat: get users & moderator user ban workflow (#90) (4dc5128)
+
+---
+
+# [0.0.26](https://github.com/juztadrop/juztadrop/compare/v0.0.25...v0.0.26) (2026-03-11)
+
+### Changes
+
+- feat: fetch volunteer by userID (#88) (9bbbcae)
+
+---
+
 # [0.0.25](https://github.com/juztadrop/juztadrop/compare/v0.0.24...v0.0.25) (2026-03-03)
 
 ### Changes